requirements.txt
jinja2~=3.1
pathvalidate~=3.2
pydantic~=2.10
pydantic-settings~=2.7
python-telegram-bot~=21.9
shortuuid~=1.0
tomlkit~=0.13